ScalaCompilerForUnitTesting: do not use structural types
Otherwise the test won't compile with dotty.
